// src/config/schema.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Generator config
 */
export interface Config {
  archive: string;
  contracts: {
    name: string;
    address: string;
    proxy?: string;
    abi?: string;
    events?: string[] | boolean;
    functions?: string[] | boolean;
  }[];
  etherscanApi?: string;
}